Databases/PostgreSQL/orms/python-postgresql/pyproject.toml
[project]
name = "python-postgresql"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.12"
dependencies = [
    "click>=8.2.1",
    "psycopg2>=2.9.10",
    "python-dotenv>=1.1.0",
    "sqlmodel>=0.0.24",
]
